// nativescript-pediatric/app/enum.ts
export enum COMMAND {
    CMD_SET_SPEED = 0,
    CMD_CHANGE_SYSTEM_MODE = 1,
    CMD_MOVE_ACTUATOR = 2,
    CMD_PHONE_JOYX = 3,
    CMD_PHONE_JOYY = 4,
    CMD_DRIVE_SPEED = 5
};
export enum COMMAND_STR {
    CMD_SET_SPEED_STR = "0x00",
    CMD_CHANGE_SYSTEM_MODE_STR = "0x01",
    CMD_MOVE_ACTUATOR_STR = "0x02",
    CMD_PHONE_JOYX_STR = "0x03",
    CMD_PHONE_JOYY_STR = "0x04",
    CMD_DRIVE_SPEED_STR = "0x05"
};
export enum SYSTEM_MODES {
    DRIVEMODE,
    ACTUATOR_RECLINE,
    ACTUATOR_LEGREST,
    ACTUATOR_TILT,
    ACTUATOR_ELEVATION,
    ACTUATOR_STAND,
    SYSTEM_SLEEP,
    PHONE_CONTROL_MODE,
    ERRORMODE,
    SYSTEM_MODE_MAX
}
export enum SYSTEM_MODES_STR {
    DRIVEMODE_STR = "0x00",
    ACTUATOR_RECLINE_STR = "0x01",
    ACTUATOR_LEGREST_STR = "0x02",
    ACTUATOR_TILT_STR = "0x03",
    ACTUATOR_ELEVATION_STR = "0x04",
    ACTUATOR_STAND_STR = "0x05",
    SYSTEM_SLEEP_STR = "0x06",
    PHONE_CONTROL_MODE_STR = "0x07",
    ERRORMODE_STR = "0x08",
    SYSTEM_MODE_MAX_STR = "0x09"
}
export enum ACTUATOR_MOVING_DIR {
    AMD_LEFT,
    AMD_RIGHT,
    AMD_STOP
}
export enum ACTUATOR_MOVING_DIR_STR {
    AMD_LEFT_STR = "0x00",
    AMD_RIGHT_STR = "0x01",
    AMD_STOP_STR = "0x02",
}
export enum SPEED_SETTING {
    SPEED_LOW,
    SPEED_MEDIUM,
    SPEED_HIGH
}
export enum SPEED_SETTING_STR {
    SPEED_LOW_STR = "0x00",
    SPEED_MEDIUM_STR = "0x01",
    SPEED_HIGH_STR = "0x02"
}